add user defaults and animations etc

+++ TodoList/Views/ListView.swift
@@ -13,12 +13,24 @@ struct ListView: View {
     @State var showSheet: Bool  = false
     
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
+        ZStack {
+            if listViewModel.items.isEmpty {
+                NoItemsView(showSheet: $showSheet)
+                    .transition(AnyTransition.opacity.animation(.easeIn))
+            } else {
+                List {
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
+                    }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
         }
         .navigationTitle("ToDo List 📋")
         .navigationBarItems(
